LPS-94013 Remove related test class into OpenSearchRegistryUtilTest as inner class
